1000 Genomes examples
It feels like I'm getting to the home stretch content-wise. I benefit a lot from creating functions for diagrams that get used regularly. I should probably do this even more, but with only a few slides left, I think it's in a good spot for now. Added 1KGP comparisons (they are not linked yet to anything).